[IMP] crm_claim_code: Functions from init was move to hooks
To clear init file
